+++ lib/src/buttons/basic_dropdown_button.dart
@@ -53,6 +53,8 @@ class BasicDropdownButton<T> extends BaseDropdownButton<T> {
     super.minWidth,
     super.theme,
     super.enabled = true,
+    super.scrollToSelectedItem = true,
+    super.scrollToSelectedDuration,
   });
 
   /// The list of items to display in the dropdown.